Update README to be compatible with webpack 2
In webpack 2, the full name of every loader has to be supplied by default. Ei. it is not possible to use `react-svg` as the loader name, it has to be `react-svg-loader`.

Currently the README included some examples that didn't work when copy/pasted into webpack 2 config.

This commit updates the README to use the `-loader` suffix in **all** of the examples — including webpack v1. This was done because in some cases it is not clear whether the example is webpack 1 or 2. webpack 1 works **with** the suffix as well so I thought this is the least confusing way to solve this. (Didn't wanna duplicate all examples etc...)

A small tweak was also made to the webpack 2 config example, as it was using `loaders` instead of `use` property, which is no longer the canonical way.